requirements.txt
python-telegram-bot==20.3
yfinance==0.2.36